[RTD-1542] move setupenv.js into common directory
